前台课程，章节，视频列表，以及视频播放器

+++ server/nov-starter/nov-service/service_edu/src/main/java/com/novedu/nov/edu/entity/dto/EduCourseInfoDTO.java
@@ -59,7 +59,7 @@ public class EduCourseInfoDTO {
     private String cover;
 
     @ApiModelProperty(value = "销售数量")
-    private Long buyCount;
+    private Integer buyCount;
 
     @ApiModelProperty(value = "浏览数量")
     private Long viewCount;
